ml-service/requirements.txt
torch==2.6.0
torchvision==0.21.0
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
Pillow==11.2.1
numpy==2.2.5
